requirements.txt
torch>=1.12.0
torch-geometric>=2.1.0
networkx>=2.8
numpy>=1.21.0
pandas>=1.3.0
matplotlib>=3.5.0
pyyaml>=6.0
scikit-learn>=1.0.0
tqdm>=4.62.0
tensorboard>=2.10.0
graphviz>=0.20
pydot>=1.4.2
pytest>=7.0.0
